Xml reader and writer (#93)
* implement basic features of xml writer with pugixml.

* improve structure of xml writer

* add control point variable values for xml writer.

* add simple tests for xml writer.

* add xml writer for nurbs with weights.

* add tests for creation of weights.

* add more tests for different space and spline dimensions.

* correct small cpplint issues.

* delete redundant line.

* move pugixml files in external directory and create pugixml library.

* exclude external directory from oclint check.

* adjust oclint dockerfile.

* adjust oclint dockerfile to exclude external files.

* adjust oclint dockerfile to exclude external files.

* create bspline from xml document.

* add weights for xml reader.

* split xml generator in reader and writer, add io directory

* Add configure file to use path to xml file in tests.

* fix cpplint issues.

* use std::any object as return value for xml reader.

* adjust header guard style.

* change return type of xml_reader::ReadXMLFile to a vector of any objects.

* split functions of xml reader into smaller functions.

* rename variables.

* enable xml reader to read multiple splines of one xml file.

* create util class for string operations.

* split test cases for xml reader.

* add test to check if spline is the same after writing and reading xml document.

* change xml writer constructor arguments from parameter space and physical space to a spline.

* change xml writer input argument to a vector of splines.

* add test to write two splines into one xml file.

* fix not working test.

* delete redundant cc-file.

* correct mistake in xml test file.

* comment not working test.

* add test cases to increase code coverage.

* replace second spline in xml reader test file so that all tests are working now.

* Add codecov yml to ignore external folder.

* Remove codecov yml.

* Ignore external folder in code coverage check.

* Force codecov to run again.

* Update CMakeLists.txt

fix line break error

* Starting to edit all knot vectors to shared pointers.

* Starting to edit all knot vectors to shared pointers.

* fix all merging errors.

* fix clang-tidy-errors.

* fix clang-tidy-errors and exclude external folder from clang-tidy-checks.

* remove redundant "virtual"

+++ src/spl/parameter_space.h
@@ -37,7 +37,7 @@ class ParameterSpace {
     baf::BasisFunctionFactory factory;
     for (int i = 0; i < DIM; i++) {
       basis_functions_[i].reserve(knot_vector_[i]->GetNumberOfKnots() - degree_[i].get() - 1);
-      for (uint64_t j = 0; j < (knot_vector_[i]->GetNumberOfKnots() - degree_[i].get() - 1); ++j) {
+      for (int j = 0; j < (static_cast<int>(knot_vector_[i]->GetNumberOfKnots()) - degree_[i].get() - 1); ++j) {
         basis_functions_[i].emplace_back(factory.CreateDynamic(*(knot_vector_[i]), KnotSpan{j}, degree_[i]));
       }
     }
